llm-gateway/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.25.1
nltk==3.7
scikit-learn==0.24.2
numpy==1.19.5
prometheus_client==0.16.0
llama-cpp-python==0.2.67